+++ src/main/java/newsite/controller/ModeratorController.java
@@ -3,18 +3,9 @@ package newsite.controller;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import javax.servlet.http.HttpSession;
 import org.springframework.transaction.annotation.Transactional;
-import newsite.repository.CategoryRepository;
-import newsite.repository.NewsRepository;
-import newsite.repository.PhotoRepository;
-import newsite.repository.ViewRepository;
-import newsite.repository.WriterRepository;
-import newsite.service.CategoryService;
 import newsite.service.ModeratorService;
-import newsite.service.NewsService;
-import newsite.service.PhotoService;
-import newsite.service.WriterService;
-import newsite.validors.NewsValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -28,6 +19,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 /**
  * Moderator controller for moderating News.
+ *
  * @author twviiala
  */
 @Transactional
@@ -39,32 +31,41 @@ public class ModeratorController {
 
     /**
      * Get moderate page.
+     *
      * @param model a
+     * @param session a
      * @return a
      */
     @GetMapping("/moderate")
-    public String list(Model model) {
+    public String list(HttpSession session, Model model) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         moderatorService.setModelList(model);
         return "moderate";
     }
 
     /**
      * Add new News.
+     *
      * @param redirectModel a
      * @param header a
      * @param lead a
      * @param text a
      * @param writers a
      * @param categories a
-     * @param photo a 
+     * @param photo a
+     * @param session a
      * @return a
      * @throws IOException a
      */
     @PostMapping("/news/moderate")
-    public String postAdd(RedirectAttributes redirectModel, @RequestParam(value = "header", required = false) String header, @RequestParam(value = "lead", required = false) String lead,
+    public String postAdd(HttpSession session, RedirectAttributes redirectModel, @RequestParam(value = "header", required = false) String header, @RequestParam(value = "lead", required = false) String lead,
             @RequestParam(value = "text", required = false) String text, @RequestParam(value = "writers", required = false) List<Long> writers,
             @RequestParam(value = "categories", required = false) List<Long> categories, @RequestParam(value = "photo", required = false) MultipartFile photo) throws IOException {
-
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         List<String> errors = moderatorService.addNews(header, lead, text, writers, categories, photo);
         if (!errors.isEmpty()) {
             redirectModel.addFlashAttribute("messages", errors);
@@ -80,12 +81,17 @@ public class ModeratorController {
 
     /**
      * Delete parameter News.
+     *
      * @param redirectModel a
      * @param id a
+     * @param session a
      * @return a
      */
     @DeleteMapping("/news/{id}")
-    public String delete(RedirectAttributes redirectModel, @PathVariable Long id) {
+    public String delete(HttpSession session, RedirectAttributes redirectModel, @PathVariable Long id) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         String header = moderatorService.deleteNews(id);
 
         List<String> messages = new ArrayList();
@@ -97,18 +103,24 @@ public class ModeratorController {
 
     /**
      * Get modifying page, parameter id News.
+     *
      * @param model a
      * @param id a
+     * @param session a
      * @return a
      */
     @GetMapping("/news/{id}/modify")
-    public String modify(Model model, @PathVariable Long id) {
+    public String modify(HttpSession session, Model model, @PathVariable Long id) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         moderatorService.setModifyModel(model, id);
         return "modify";
     }
 
     /**
      * Modify parameter id news with other parameters.
+     *
      * @param redirectModel a
      * @param id a
      * @param header a
@@ -117,13 +129,17 @@ public class ModeratorController {
      * @param writers a
      * @param categories a
      * @param photo a
+     * @param session a
      * @return a
      * @throws IOException a
      */
     @PostMapping("/moderator/news/{id}")
-    public String postModify(RedirectAttributes redirectModel, @PathVariable Long id, @RequestParam(value = "header", required = false) String header, @RequestParam(value = "lead", required = false) String lead,
+    public String postModify(HttpSession session, RedirectAttributes redirectModel, @PathVariable Long id, @RequestParam(value = "header", required = false) String header, @RequestParam(value = "lead", required = false) String lead,
             @RequestParam(value = "text", required = false) String text, @RequestParam(value = "writers", required = false) List<Long> writers,
             @RequestParam(value = "categories", required = false) List<Long> categories, @RequestParam(value = "photo", required = false) MultipartFile photo) throws IOException {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         List<String> errors = moderatorService.modifyNews(id, header, lead, photo, text, writers, categories);
         if (!errors.isEmpty()) {
             redirectModel.addFlashAttribute("messages", errors);
