+++ src/main/java/newsite/controller/WriterController.java
@@ -2,6 +2,7 @@ package newsite.controller;
 
 import java.util.ArrayList;
 import java.util.List;
+import javax.servlet.http.HttpSession;
 import org.springframework.transaction.annotation.Transactional;
 import newsite.domain.Writer;
 import newsite.repository.NewsRepository;
@@ -20,81 +21,100 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 /**
  * Writer Controller.
+ *
  * @author twviiala
  */
 @Transactional
 @Controller
 public class WriterController {
-    
+
     @Autowired
     private WriterService writerService;
-    
+
     /**
      * Add a new writer.
+     *
      * @param redirectModel a
      * @param name a
+     * @param session a
      * @return a
      */
     @PostMapping("/writers")
-    public String add(RedirectAttributes redirectModel, @RequestParam String name) {
-        
+    public String add(HttpSession session, RedirectAttributes redirectModel, @RequestParam String name) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         List<String> errors = writerService.addWriter(name);
         if (!errors.isEmpty()) {
             redirectModel.addFlashAttribute("messages", errors);
             return "redirect:/moderate";
         }
-        
+
         List<String> messages = new ArrayList();
         messages.add("Has  " + name + " written something interesting?");
         redirectModel.addFlashAttribute("messages", messages);
-        
+
         return "redirect:/moderate";
     }
-    
+
     /**
      * Delete parameter id Writer.
+     *
      * @param redirectModel a
      * @param id a
+     * @param session a
      * @return a
      */
     @DeleteMapping("/writers/{id}")
-    public String delete(RedirectAttributes redirectModel, @PathVariable Long id) {
-        
+    public String delete(HttpSession session, RedirectAttributes redirectModel, @PathVariable Long id) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         String name = writerService.deleteWriter(id);
-        
+
         List<String> messages = new ArrayList();
         messages.add("It seems  " + name + " has stopped writing. Maybe " + name + " has never wrote anything!");
         redirectModel.addFlashAttribute("messages", messages);
         return "redirect:/moderate";
     }
-    
+
     /**
      * Get modifyWriter page.
+     *
      * @param model a
      * @param id a
+     * @param session a
      * @return a
      */
     @GetMapping("/writers/{id}/modify")
-    public String modify(Model model, @PathVariable Long id) {
+    public String modify(HttpSession session, Model model, @PathVariable Long id) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         writerService.setModifyModel(model, id);
         return "modifyWriter";
     }
-    
+
     /**
      * Modify parameter id Writer.
+     *
      * @param redirectModel a
-     * @param id a
+     * @param id writer
      * @param name a
+     * @param session a
      * @return a
      */
     @PostMapping("/moderator/writers/{id}")
-    public String postModify(RedirectAttributes redirectModel, @PathVariable Long id, @RequestParam String name) {
+    public String postModify(HttpSession session, RedirectAttributes redirectModel, @PathVariable Long id, @RequestParam String name) {
+        if (session.getAttribute("moderator") == null) {
+            return "redirect:/";
+        }
         List<String> errors = writerService.modifyWriter(name, id);
         if (!errors.isEmpty()) {
             redirectModel.addFlashAttribute("messages", errors);
             return "redirect:/moderate";
         }
-        
+
         List<String> messages = new ArrayList();
         messages.add("It seems  " + name + " might be an alias for a writer.");
         redirectModel.addFlashAttribute("messages", messages);
